requirements.txt
alembic==1.7.7
attrs==21.4.0
autopep8==1.6.0
blinker==1.6.2
cachelib==0.10.2
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
Flask==2.3.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
flask-session==0.5.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.12.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.5
py==1.11.0
pycodestyle==2.8.0
pyparsing==3.0.7
pytest==7.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-editor==1.0.4
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.35
toml==0.10.2
tomli==2.0.1
urllib3==1.26.9
Werkzeug==2.3.6
zipp==3.7.0
